+++ src/main/java/edu/cmu_ucla/minuet/model/VitalObjectCommand.java
@@ -2,7 +2,7 @@ package edu.cmu_ucla.minuet.model;
 
 public interface VitalObjectCommand {
 
-    String execute(int a);
+    String execute(double a);
 
     String getCommandTopic();
 }
